:speech_balloon: include threaded comments in HTML
This update recursively reads through the 'replies' field of each
post and adds them to the HTML output. Since this process is
recursive, it handles nested comments quite easily.

The inelegant part of this process is that, in order to output
participant names, I had to pass around a dictionary lookup
table of IDs to names. So this could probably be made more
efficient. But it works and the output looks decent.